examples/pyproject.toml
[project]
name = "snowflake-py-logs-examples"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "snowflake-snowpark-python",
    "snowflake-ml-python",
]
